+++ Assets/Editor/Tests/MathTests.cs
@@ -16,9 +16,11 @@ namespace BeauUtil.UnitTests
     static public class MathTests
     {
         [Test]
-        static public void CanDecomposeTRS() {
+        static public void CanDecomposeTRS()
+        {
             int count = 1000;
-            while (count-- > 0) {
+            while (count-- > 0)
+            {
                 TRS trs = new TRS(RNG.Instance.NextVector3(), Quaternion.Euler(RNG.Instance.NextVector3(0, 360)), RNG.Instance.NextVector3(1, 92) * RNG.Instance.NextFloat(0, 1));
                 Matrix4x4 mat = trs.Matrix;
                 bool check = TRS.TryCreateFromMatrix(mat, out var newTRS);
@@ -28,7 +30,8 @@ namespace BeauUtil.UnitTests
         }
 
         [Test]
-        static public void CanComputeQuaternionFastAxis() {
+        static public void CanComputeQuaternionFastAxis()
+        {
             int iterCount = ushort.MaxValue;
 
             Quaternion q;
@@ -37,50 +40,63 @@ namespace BeauUtil.UnitTests
             Vector3 cachedU = Vector3.up;
             Vector3 cachedR = Vector3.right;
 
-            for(int i = 0; i < iterCount; i++) {
+            for (int i = 0; i < iterCount; i++)
+            {
                 q = Quaternion.Euler(i * 4, i * -54 + 30, i * 13 - 24);
                 v = q * cachedF;
                 v2 = Geom.Forward(q);
                 Assert.IsTrue(v == v2, "Fast quaternion axis is not accurate");
             }
 
-            using (Profiling.AvgTime("forward (mult)", iterCount, ProfileTimeUnits.Cycles)) {
-                for (int i = 0; i < iterCount; i++) {
+            using (Profiling.AvgTime("forward (mult)", iterCount, ProfileTimeUnits.Cycles))
+            {
+                for (int i = 0; i < iterCount; i++)
+                {
                     q = Quaternion.Euler(i * 4, i * -7 + 30, i * 13 - 24);
                     v = q * cachedF;
                 }
             }
 
-            using (Profiling.AvgTime("forward (fast)", iterCount, ProfileTimeUnits.Cycles)) {
-                for (int i = 0; i < iterCount; i++) {
+            using (Profiling.AvgTime("forward (fast)", iterCount, ProfileTimeUnits.Cycles))
+            {
+                for (int i = 0; i < iterCount; i++)
+                {
                     q = Quaternion.Euler(i * 4, i * -7 + 30, i * 13 - 24);
                     v = Geom.Forward(q);
                 }
             }
 
-            using (Profiling.AvgTime("up (mult)", iterCount, ProfileTimeUnits.Cycles)) {
-                for (int i = 0; i < iterCount; i++) {
+            using (Profiling.AvgTime("up (mult)", iterCount, ProfileTimeUnits.Cycles))
+            {
+                for (int i = 0; i < iterCount; i++)
+                {
                     q = Quaternion.Euler(i * 4, i * -7 + 30, i * 13 - 24);
                     v = q * cachedU;
                 }
             }
 
-            using (Profiling.AvgTime("up (fast)", iterCount, ProfileTimeUnits.Cycles)) {
-                for (int i = 0; i < iterCount; i++) {
+            using (Profiling.AvgTime("up (fast)", iterCount, ProfileTimeUnits.Cycles))
+            {
+                for (int i = 0; i < iterCount; i++)
+                {
                     q = Quaternion.Euler(i * 4, i * -7 + 30, i * 13 - 24);
                     v = Geom.Up(q);
                 }
             }
 
-            using (Profiling.AvgTime("right (mult)", iterCount, ProfileTimeUnits.Cycles)) {
-                for (int i = 0; i < iterCount; i++) {
+            using (Profiling.AvgTime("right (mult)", iterCount, ProfileTimeUnits.Cycles))
+            {
+                for (int i = 0; i < iterCount; i++)
+                {
                     q = Quaternion.Euler(i * 4, i * -7 + 30, i * 13 - 24);
                     v = q * cachedR;
                 }
             }
 
-            using (Profiling.AvgTime("right (fast)", iterCount, ProfileTimeUnits.Cycles)) {
-                for (int i = 0; i < iterCount; i++) {
+            using (Profiling.AvgTime("right (fast)", iterCount, ProfileTimeUnits.Cycles))
+            {
+                for (int i = 0; i < iterCount; i++)
+                {
                     q = Quaternion.Euler(i * 4, i * -7 + 30, i * 13 - 24);
                     v = Geom.Right(q);
                 }
@@ -88,7 +104,8 @@ namespace BeauUtil.UnitTests
         }
 
         [Test]
-        static public void CanSetMatrixTranslation() {
+        static public void CanSetMatrixTranslation()
+        {
             int iterCount = ushort.MaxValue;
 
             Quaternion rotation;
@@ -101,42 +118,58 @@ namespace BeauUtil.UnitTests
             Matrix4x4 cachedOriginalMatrix = Matrix4x4.TRS(default(Vector3), rotation, scale);
             Matrix4x4 freshMatrix;
 
-            for (int i = 0; i < iterCount; i++) {
+            for (int i = 0; i < iterCount; i++)
+            {
                 translation = new Vector3(i * 4, i * -7 + 30, i * 13 - 24);
                 freshMatrix = Matrix4x4.TRS(translation, rotation, scale);
                 Geom.SetTranslation(ref cachedOriginalMatrix, translation);
                 Assert.IsTrue(freshMatrix == cachedOriginalMatrix, "Setting translation is not accurate");
             }
 
-            using (Profiling.AvgTime("trs", iterCount, ProfileTimeUnits.Cycles)) {
-                for (int i = 0; i < iterCount; i++) {
+            using (Profiling.AvgTime("trs", iterCount, ProfileTimeUnits.Cycles))
+            {
+                for (int i = 0; i < iterCount; i++)
+                {
                     translation = new Vector3(i * 4, i * -7 + 30, i * 13 - 24);
                     freshMatrix = Matrix4x4.TRS(translation, rotation, scale);
                 }
             }
 
-            using (Profiling.AvgTime("matrix mult", iterCount, ProfileTimeUnits.Cycles)) {
-                for (int i = 0; i < iterCount; i++) {
+            using (Profiling.AvgTime("matrix mult", iterCount, ProfileTimeUnits.Cycles))
+            {
+                for (int i = 0; i < iterCount; i++)
+                {
                     translation = new Vector3(i * 4, i * -7 + 30, i * 13 - 24);
                     freshMatrix = cachedOriginalMatrix * Matrix4x4.Translate(translation);
                 }
             }
 
-            using (Profiling.AvgTime("set translation", iterCount, ProfileTimeUnits.Cycles)) {
-                for (int i = 0; i < iterCount; i++) {
+            using (Profiling.AvgTime("set translation", iterCount, ProfileTimeUnits.Cycles))
+            {
+                for (int i = 0; i < iterCount; i++)
+                {
                     translation = new Vector3(i * 4, i * -7 + 30, i * 13 - 24);
                     Geom.SetTranslation(ref cachedOriginalMatrix, translation);
                 }
             }
 
-            using (Profiling.AvgTime("set translation (simple)", iterCount, ProfileTimeUnits.Cycles)) {
-                for (int i = 0; i < iterCount; i++) {
+            using (Profiling.AvgTime("set translation (simple)", iterCount, ProfileTimeUnits.Cycles))
+            {
+                for (int i = 0; i < iterCount; i++)
+                {
                     translation = new Vector3(i * 4, i * -7 + 30, i * 13 - 24);
                     cachedOriginalMatrix.m03 = translation.x;
                     cachedOriginalMatrix.m13 = translation.y;
                     cachedOriginalMatrix.m23 = translation.z;
                 }
             }
+        }
+
+        [Test]
+        static public void CanWrapValues()
+        {
+            Assert.AreEqual(-180, MathUtils.Wrap(180, -180, 180));
+            Assert.AreEqual(18, MathUtils.Wrap(180, -1, 80));
         }
     }
 }